Move type declarations out into a separate file
