+++ src/containers/App/App.tsx
@@ -7,35 +7,7 @@ import Check from '../../components/Check/Check'
 import Result from '../../components/Result/Result'
 import Score from '../../components/Score/Score'
 import Splashscreen from '../../components/Splashscreen/Splashscreen';
-
-export type GameStates = 'startGame' | 'showSum' | 'confirmAnswer' | 'showResult' | 'endWin' | 'endLose';
-
-export type AnswerButton = number | '?';
-
-export interface AppState {
-  showSplash: boolean;
-  gameStatus: GameStates;
-  possibleNums: number[];
-  baseNum: number;
-  num1: number | null;
-  num2: AnswerButton;
-  op1: string;
-  op2: string;
-  possibleAns: number[];
-  correctAns: number | null;
-  gotItRight: boolean | null;
-  score: number;
-  totalLives: number;
-  livesLeft: number;
-}
-
-interface AnswerMethod {
-  (a: number, b: number): number;
-}
-
-interface AnswerMethodsObj {
-  [key: string]: AnswerMethod;
-}
+import { AppState, AnswerMethodsObj } from '../../types/types';
 
 class App extends Component<{}, AppState> {
